Add path for readthedocs
