+++ pyproject.toml
@@ -14,6 +14,8 @@ psycopg2-binary = "^2.9.5"
 sqlalchemy-utils = "^0.38.3"
 mimesis = "^6.1.1"
 typer = "^0.7.0"
+python-docs-theme = "^2022.1"
+sphinx-rtd-theme = "^1.1.1"
 
 
 [tool.poetry.group.dev.dependencies]
